# openapi.yaml
openapi: 3.0.1
info:
  title: Anvil Web API
  description: |-
    This document describes the Anvil Web API. It is used to communicate between AnvilWorkers, the AnvilWeb UI and the AnvilWeb backend.

    AnvilWorkers need to register at the backend to get an id. After that, they have to poll the fetch endpoint regularly to get new commands.
  version: v2
...
servers:
  - url: https://localhost:5001/api/v2
...
paths:
...
  /worker/register:
    post:
      tags:
        - worker
      summary: Register a new worker
      description: Register as a new worker for this backend
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties: 
                name:
                  type: string
      responses: 
        '200':
          description: successful operation
          content:
            application/json: 
              schema:
                type: object
                $ref: '#/components/schemas/AnvilWorker'
  /worker/fetch:
    post:
      tags:
        - worker
      summary: Fetch next command
      description: Fetch next command for a worker (polling)
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties: 
                id:
                  type: string
                  description: worker id
                status:
                  type: string
                  description: staus of the worker
      responses: 
        '200':
          description: successful operation
          content:
            application/json: 
              schema:
                type: object
                $ref: '#/components/schemas/AnvilCommand'
  /worker/update/scan:
    post:
      tags:
        - worker
      summary: Signal, that the worker started scanning
      description: Signal, that the worker is now doing a feature extraction scan
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                jobId:
                  type: string
      responses: 
        '200':
          description: successful operation
          content:
            application/json: 
              schema:
                type: object
                properties: 
                  status:
                    type: string
                    default: "OK"
  /worker/update/report:
    post:
      tags:
        - worker
      summary: Update a test report
      description: Create or update a test report for the given identifier
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                jobId:
                  type: string
                report:
                  $ref: '#/components/schemas/Report'
      responses: 
        '200':
          description: successful operation
          content:
            application/json: 
              schema:
                type: object
                properties: 
                  status:
                    type: string
                    default: "OK"
  /worker/update/testrun:
    post:
      tags:
        - worker
      summary: Update a testrun
      description: Create or update a testrun for the given job
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                jobId:
                  type: string
                report:
                  $ref: '#/components/schemas/TestRun'
      responses: 
        '200':
          description: successful operation
          content:
            application/json: 
              schema:
                type: object
                properties: 
                  status:
                    type: string
                    default: "OK"
  /worker/update/testcase:
    post:
      tags:
        - worker
      summary: Append test case
      description: Append a test case to an existing testrun
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                jobId:
                  type: string
                className:
                  type: string
                methodName:
                  type: string
                testcase:
                  $ref: '#/components/schemas/TestCase'
      responses: 
        '200':
          description: successful operation
          content:
            application/json: 
              schema:
                type: object
                properties: 
                  status:
                    type: string
                    default: "OK"
...
components:
  schemas:
    AnvilCommand:
      type: object
      properties: 
        command:
          type: string
          enum:
            - "QUEUE_RUN"
            - "STOP_RUN"
            - "RESTART_RUN"
            - "RESTART_TEST"
            - "RESTART_STATE"
            - "SHUTDOWN"
            - "OK"
        job:
          $ref: '#/components/schemas/AnvilJob'
        jobId:
          type: string
        className:
          type: string
        methodName:
          type: string
        uuid:
          type: string
    AnvilJob:
      type: object
      properties: 
        id: 
          type: string
        status:
          type: string
          enum:
            - "QUEUED"
            - "SCANNING"
            - "TESTING"
            - "CANCELED"
        progress:
          type: integer
        identifier:
          $ref: '#/components/schemas/Identifier'
        config:
          type: object
        additionalConfig:
          type: object
        workerId:
          type: string
        workerName:
          type: string
    AnvilWorker:
      type: object
      properties: 
        name:
          type: string
        id:
          type: string
        status:
          type: string
          enum:
            - "WORKING"
            - "IDLE"
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/AnvilJob'
...
    Identifier:
      type: string
      examples: ["openssl-server", "matrixssl-client"]
    Report:
      type: object
      properties:
        Identifier:
          $ref: '#/components/schemas/Identifier'
        Date:
          type: string
          format: date-time
          examples: ["2022-05-22T16:07:30.361Z"]
        DisplayName:
          type: string
          examples: ["openssl-server"]
        Running:
          type: boolean
        ElapsedTime:
          type: integer
          examples: [12345]
        FailedTests:
          type: integer
          examples: [10]
        SucceededTests:
          type: integer
          examples: [10]
        DisabledTests:
          type: integer
          examples: [10]
        TotalTests:
          type: integer
          examples: [10]
        FinishedTests:
          type: integer
          examples: [10]
        StatesCount:
          type: integer
          examples: [10]
        Score:
          $ref: '#/components/schemas/ScoreMap'
        TestEndpointType:
          type: string
          enum: 
            - "server"
            - "client"
        Config:
          type: string
    ScoreCategory:
      type: string
      enum: 
        - "ALERT"
        - "CVE"
        - "CERTIFICATE"
        - "CRYPTO"
        - "DEPRECATED"
        - "HANDSHAKE"
        - "MESSAGESTRUCTURE"
        - "RECORDLAYER"
        - "SECURITY"
        - "INTEROPERABILITY"
        - "COMPLIANCE"
    ScoreMap:
      type: object
      propertyNames: 
        $ref: '#/components/schemas/ScoreCategory'
      additionalProperties: 
        $ref: '#/components/schemas/Score'
    Score:
      type: object
      properties: 
        Total:
          type: number
        Reached:
          type: number
        Percentage:
          type: number
        SeverityLevel:
          $ref: '#/components/schemas/SeverityLevel'
    SeverityLevel:
      type: string
      enum:
        - "INFORMATIONAL"
        - "LOW"
        - "MEDIUM"
        - "HIGH"
        - "CRITICAL"
    TestMethod:
      type: object
      properties: 
        Description:
          type: string
        TestDescription:
          type: string
        TlsVersion:
          type: string
          examples: ["TLS 1.3"]
        RFC:
          type: object
          properties:
            Section:
              type: string
            number:
              type: integer
        MethodName:
          type: string
        DisplayName:
          type: string
        ClassName:
          type: string
    TestResult:
      type: string
      enum:
        - "STRICTLY_SUCCEEDED"
        - "FULLY_FAILED"
        - "PARTIALLY_FAILED"
        - "CONCEPTUALLY_SUCCEEDED"
        - "DISABLED"
        - "PARSER_ERROR"
        - "NOT_SPECIFIED"
        - "INCOMPLETE"
    TestRun:
      type: object
      properties: 
        TestMethod:
          $ref: '#/components/schemas/TestMethod'
        Result:
          $ref: '#/components/schemas/TestResult'
        HasStateWithAdditionalResultInformation:
          type: boolean
        HasVaryingAdditionalResultInformation:
          type: boolean
        DisabledReason:
          type: string
        FailedReason:
          type: string
        FailedStacktrace:
          type: string
        ElapsedTime:
          type: integer
        TestCases:
          type: array
          items: 
            $ref: '#/components/schemas/TestCase'
        CaseCount:
          type: integer
        SucceededCases:
          type: integer
        ConSucceededCases:
          type: integer
        FailedCases:
          type: integer
        Score:
          $ref: '#/components/schemas/ScoreMap'
        FailureInducingCombinations:
          type: object
          additionalProperties: 
            type: string
    TestCase:
      type: object
      properties: 
        DerivationContainer:
          type: object
          additionalProperties: 
            type: string
        DisplayName:
          type: string
        Result:
          $ref: '#/components/schemas/TestResult'
        AdditionalResultInformation:
          type: string
        AdditionalTestInformation:
          type: string
        SrcPort:
          type: integer
        DstPort:
          type: integer
        StartTimestamp:
          type: string
          format: date-time
        EndTimestamp:
          type: string
          format: date-time
        uuid:
          type: string
        Stacktrace:
          type: string